% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CachePath}
\alias{CachePath}
\title{Path to cache}
\usage{
CachePath(...)
}
\arguments{
\item{...}{path components (character vector)}
}
\description{
Path to cache
}
\examples{
CachePath("mysubfolder", "myfile.txt")

}
