% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{CreateFolders}
\alias{CreateFolders}
\title{Create folder structure for your project}
\usage{
CreateFolders(force = FALSE)
}
\arguments{
\item{force}{re-write README.md files in the directories}
}
\description{
Create folder structure for your project
}
\examples{
CreateFolders(force=F)

}
