% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DataPath}
\alias{DataPath}
\title{Path to data}
\usage{
DataPath(...)
}
\arguments{
\item{...}{path components (character vector)}
}
\description{
Path to data
}
\examples{
DataPath("mysubfolder", "myfile.txt")

}
