% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{DatasetPath}
\alias{DatasetPath}
\title{Path to dataset, specified in the data_mapping.yml}
\usage{
DatasetPath(dataset.name, ..., mapping.file = NULL)
}
\arguments{
\item{dataset.name}{name of the dataset}

\item{...}{path components (character vector)}
}
\description{
Path to dataset, specified in the data_mapping.yml
}
\examples{
DatasetPath("my_dataset", "mysubfolder", "myfile.txt")

}
