% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{MetadataPath}
\alias{MetadataPath}
\title{Path to meta-data}
\usage{
MetadataPath(...)
}
\arguments{
\item{...}{path components (character vector)}
}
\description{
Path to meta-data
}
\examples{
MetadataPath("mysubfolder", "myfile.txt")

}
