% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{OutputPath}
\alias{OutputPath}
\title{Path to output}
\usage{
OutputPath(...)
}
\arguments{
\item{...}{path components (character vector)}
}
\description{
Path to output
}
\examples{
OutputPath("mysubfolder", "myfile.txt")

}
