% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ReadOrCreate}
\alias{ReadOrCreate}
\title{Read RDS file or create it if it does not exist using `create.fun()`}
\usage{
ReadOrCreate(path, create.fun, force = FALSE)
}
\arguments{
\item{path}{path to the cached file}

\item{create.fun}{function to create the object if it does not exist}

\item{force}{force re-creating the object ignoring the existing one}
}
\description{
Read RDS file or create it if it does not exist using `create.fun()`
}
